requirements.txt
streamlit>=1.32.0
pandas
PyPDF2
openpyxl
faiss-cpu
transformers
sentence-transformers
langchain
langchain-community
python-docx
langchain-huggingface
hf_xet
llama-index-core==0.10.30
llama-index-readers-file
llama-index-llms-huggingface
llama-index-embeddings-huggingface
accelerate
torch

